+++ CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG
 - Disable payment methods if they error when creating
 - Add `paymentOptionPriority` option for specifying the ordering of payment options such as `card` and `paypal`
 - Add translations
-- Use version 3.11.0 of braintree-web
+- Use version 3.11.1 of braintree-web
 
 1.0.0-beta.5
 ------------